% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack6.R
\name{cheat}
\alias{cheat}
\title{This is a function that allows you to cheat on question 1,2,6,16 of Assignment 3.2.}
\usage{
cheat(number)
}
\arguments{
\item{number}{numeric variable}
}
\description{
This is a function that allows you to cheat on question 1,2,6,16 of Assignment 3.2.
}
