% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rpack6.R
\name{remind_me}
\alias{remind_me}
\title{This is a function that reminds you of the birthday of the owner of the package.}
\usage{
remind_me()
}
\description{
This is a function that reminds you of the birthday of the owner of the package.
}
